+++ lib/src/result/release_transformer.dart
@@ -4,11 +4,10 @@
 
 import 'dart:async';
 
-import '../result.dart';
+import 'result.dart';
 import 'release_sink.dart';
 
-/// Use [Result.releaseTransformer] instead.
-@Deprecated("Will be removed in async 2.0.0.")
+/// A transformer that releases result events as data and error events.
 class ReleaseStreamTransformer<T> implements StreamTransformer<Result<T>, T> {
   const ReleaseStreamTransformer();
 
@@ -16,7 +15,6 @@ class ReleaseStreamTransformer<T> implements StreamTransformer<Result<T>, T> {
     return new Stream<T>.eventTransformed(source, _createSink);
   }
 
-  static EventSink<Result> _createSink(EventSink sink) {
-    return new ReleaseSink(sink);
-  }
+  // Since Stream.eventTransformed is not generic, this method can be static.
+  static EventSink<Result> _createSink(EventSink sink) => new ReleaseSink(sink);
 }